Moved chesum calculation into an abstract subclass of Frame
