+++ agave/frames/core.py
@@ -1,7 +1,7 @@
 from io import BytesIO
 
 
-class EndOfBufferError:
+class EndOfBufferError(Exception):
 	pass
 
 
@@ -47,10 +47,13 @@ class Buffer:
 
 	def __bytes__(self):
 		self._buf.seek(0) # use rewind instead (?)
-		return self._buf.read()
+		return self._buf.read() # use read_reamaining (?)
 
 	def rewind(self):
 		self._buf.seek(0)
+
+	def read_remaining(self):
+		return self._buf.read()
 
 
 class Frame:
@@ -61,3 +64,31 @@ class Frame:
 
 	def write_to_buffer(self, buf: Buffer):
 		raise NotImplementedError()
+
+
+class FrameWithChecksum(Frame):
+
+	def compute_checksum(self):
+		raise NotImplementedError()
+
+	def set_checksum(self):
+		self.checksum = 0
+		self.checksum = self.compute_checksum()
+
+	def compute_checksum_from_buffer(self, buf: Buffer, words: int):
+		"""
+		PARAMS
+			buf: the data required by the protocol to compute it's checksum.
+			words: the number of 16 bits words that need to be read from the buffer.
+		"""
+		csum = 0
+		for i in range(0, words):
+			csum += buf.read_short()
+		while (csum & 0xff0000) != 0:
+			carry = ( csum & 0xff0000 ) >> 16
+			csum = csum & 0x00ffff
+			csum += carry
+		return 0xffff - csum
+
+	def is_checksum_valid(self):
+		return self.compute_checksum() == 0
