+++ agave/frames/ip.py
@@ -1,4 +1,4 @@
-from .core import Frame, Buffer
+from .core import FrameWithChecksum, Buffer
 from ipaddress import ip_address
 
 
@@ -7,7 +7,7 @@ PROTO_TCP = 0x06
 PROTO_UDP = 0x11
 
 
-class IPv4(Frame):
+class IPv4(FrameWithChecksum):
 
 	def __init__(self, ihl, dscp, ecn, total_length, identification, flags,
 			fragment_offset, ttl, protocol, checksum, source, destination, options):
@@ -77,28 +77,14 @@ class IPv4(Frame):
 		buf.write(self.destination)
 		buf.write(self.options)
 
-	def set_checksum(self):
-		self.checksum = 0
-		self.checksum = self.compute_checksum()
-
 	def compute_checksum(self):
-		# Write header to buffer
+		# Writes header to buffer
 		buf = Buffer.from_bytes()
 		self.write_to_buffer(buf)
 		buf.rewind()
-		# compute the checksum
+		# Computes the checksum
 		words = self.ihl * 2
-		csum = 0
-		for i in range(0, words):
-			csum += buf.read_short()
-		while (csum & 0xff0000) != 0:
-			carry = ( csum & 0xff0000 ) >> 16
-			csum = csum & 0x00ffff
-			csum += carry
-		return 0xffff - csum
-
-	def is_checksum_valid(self):
-		return self.compute_checksum() == 0
+		return self.compute_checksum_from_buffer(buf, words)
 
 	def __str__(self):
 		return "({}) {} -> {}".format(
